requirements.txt
TestLink-API-Python-client==0.8.0
googlemaps==2.5.1